requirements.txt
asgiref==3.6.0
Babel==2.11.0
Django==4.1.5
django-crispy-forms==1.14.0
django-filter==22.1
django-money==3.0.0
django-multiselectfield==0.1.12
django-phone-field==1.8.1
mysqlclient==2.1.1
py-moneyed==2.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
